refactor: rename package cli to support
